requirements.txt
aiogram==3.11.0
openai
langchain
langchain-community
langchain-huggingface
qdrant-client
PyPDF2
unstructured[local-inference]
python-dotenv
sentence-transformers
